QuickSort still a problem

+++ include/fileHandling.h
@@ -1,8 +1,10 @@
 #ifndef fileHandling_H_
 #define fileHandling_H_
 
+#include <stdlib.h>
+
 void printHeaderToFile(char algorithm[], char inputType[]);
 
-void printResultsToFile(unsigned int n, unsigned int *op);
+void printResultsToFile(unsigned int n, size_t *op);
 
 #endif
